requirements.txt
fastmcp==0.4.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
python-dateutil==2.8.2
pytz==2024.1
